+++ apps/ledger-live-mobile/src/components/RootNavigator/PortfolioNavigator.tsx
@@ -2,6 +2,7 @@
 // @flow
 
 import React, { useMemo } from "react";
+import { useSelector } from "react-redux";
 import { createStackNavigator } from "@react-navigation/stack";
 import { useTheme } from "styled-components/native";
 import { NavigatorName, ScreenName } from "../../const";
@@ -9,8 +10,10 @@ import { NavigatorName, ScreenName } from "../../const";
 // $FlowFixMe
 // $FlowFixMe
 import Portfolio from "../../screens/Portfolio";
+import ReadOnlyPortfolio from "../../screens/Portfolio/ReadOnly";
 import AccountsNavigator from "./AccountsNavigator";
 import { getStackNavigatorConfig } from "../../navigation/navigatorConfig";
+import { readOnlyModeEnabledSelector } from "../../reducers/settings";
 
 export default function PortfolioNavigator() {
   const { colors } = useTheme();
@@ -18,6 +21,8 @@ export default function PortfolioNavigator() {
     () => getStackNavigatorConfig(colors, true),
     [colors],
   );
+  const readOnlyModeEnabled = useSelector(readOnlyModeEnabledSelector);
+
   return (
     <Stack.Navigator
       screenOptions={stackNavigationConfig}
@@ -26,7 +31,7 @@ export default function PortfolioNavigator() {
     >
       <Stack.Screen
         name={ScreenName.Portfolio}
-        component={Portfolio}
+        component={readOnlyModeEnabled ? ReadOnlyPortfolio : Portfolio}
         options={{
           headerShown: false,
         }}