+++ apps/ledger-live-mobile/src/index.js
@@ -19,11 +19,15 @@ import {
   Linking,
   Appearance,
   AppState,
+  Platform,
 } from "react-native";
 import SplashScreen from "react-native-splash-screen";
 import { SafeAreaProvider } from "react-native-safe-area-context";
 import { I18nextProvider } from "react-i18next";
-import { NavigationContainer } from "@react-navigation/native";
+import {
+  getStateFromPath,
+  NavigationContainer,
+} from "@react-navigation/native";
 import Transport from "@ledgerhq/hw-transport";
 import { NotEnoughBalance } from "@ledgerhq/errors";
 import { log } from "@ledgerhq/logs";
@@ -36,7 +40,10 @@ import { NftMetadataProvider } from "@ledgerhq/live-common/lib/nft";
 import { ToastProvider } from "@ledgerhq/live-common/lib/notifications/ToastProvider";
 import { GlobalCatalogProvider } from "@ledgerhq/live-common/lib/platform/providers/GlobalCatalogProvider";
 import { RampCatalogProvider } from "@ledgerhq/live-common/lib/platform/providers/RampCatalogProvider";
-import { RemoteLiveAppProvider } from "@ledgerhq/live-common/lib/platform/providers/RemoteLiveAppProvider";
+import {
+  RemoteLiveAppProvider,
+  useRemoteLiveAppContext,
+} from "@ledgerhq/live-common/lib/platform/providers/RemoteLiveAppProvider";
 import { LocalLiveAppProvider } from "@ledgerhq/live-common/src/platform/providers/LocalLiveAppProvider";
 
 import logger from "./logger";
@@ -92,6 +99,8 @@ import StyleProvider from "./StyleProvider";
 // $FlowFixMe
 import MarketDataProvider from "./screens/Market/MarketDataProviderWrapper";
 import AdjustProvider from "./components/AdjustProvider";
+import DelayedTrackingProvider from "./components/DelayedTrackingProvider";
+import { useFilteredManifests } from "./screens/Platform/shared";
 
 const themes = {
   light: lightTheme,
@@ -238,6 +247,12 @@ const linkingOptions = {
            * ie: "ledgerlive://wc?uri=wc:00e46b69-d0cc-4b3e-b6a2-cee442f97188@1?bridge=https%3A%2F%2Fbridge.walletconnect.org&key=91303dedf64285cbbaf9120f6e9d160a5c8aa3deb67017a3874cd272323f48ae
            */
           [ScreenName.WalletConnectDeeplinkingSelectAccount]: "wc",
+          [ScreenName.PostBuyDeviceScreen]: "hw-purchase-success",
+          /**
+           * @params ?platform: string
+           * ie: "ledgerlive://discover/paraswap?theme=light" will open the catalog and the paraswap dapp with a light theme as parameter
+           */
+          [ScreenName.PlatformApp]: "discover/:platform",
           [NavigatorName.Main]: {
             initialRouteName: ScreenName.Portfolio,
             screens: {
@@ -261,22 +276,21 @@ const linkingOptions = {
               [NavigatorName.Market]: {
                 screens: {
                   /**
-                   * @params ?platform: string
-                   * ie: "ledgerlive://discover" will open the catalog
-                   * ie: "ledgerlive://discover/paraswap?theme=light" will open the catalog and the paraswap dapp with a light theme as parameter
+                   * ie: "ledgerlive://market" will open the market screen
                    */
                   [ScreenName.MarketList]: "market",
                 },
               },
               [NavigatorName.Discover]: {
-                screens: {
-                  /**
-                   * @params ?platform: string
-                   * ie: "ledgerlive://discover" will open the catalog
-                   * ie: "ledgerlive://discover/paraswap?theme=light" will open the catalog and the paraswap dapp with a light theme as parameter
-                   */
-                  [ScreenName.PlatformCatalog]: "discover/:platform?",
-                },
+                screens:
+                  Platform.OS === "ios"
+                    ? {}
+                    : {
+                        /**
+                         * ie: "ledgerlive://discover" will open the catalog
+                         */
+                        [ScreenName.PlatformCatalog]: "discover",
+                      },
               },
               [NavigatorName.Manager]: {
                 screens: {
@@ -355,10 +369,26 @@ const linkingOptions = {
   },
 };
 
+const linkingOptionsOnboarding = {
+  ...linkingOptions,
+  config: {
+    screens: {
+      [NavigatorName.Base]: {
+        initialRouteName: NavigatorName.Main,
+        screens: {
+          [ScreenName.PostBuyDeviceScreen]: "hw-purchase-success",
+        },
+      },
+    },
+  },
+};
+
 const DeepLinkingNavigator = ({ children }: { children: React$Node }) => {
   const dispatch = useDispatch();
   const hasCompletedOnboarding = useSelector(hasCompletedOnboardingSelector);
   const wcContext = useContext(_wcContext);
+  const removeLiveAppState = useRemoteLiveAppContext();
+  const filteredManifests = useFilteredManifests();
 
   const linking = useMemo(
     () => ({
@@ -367,8 +397,33 @@ const DeepLinkingNavigator = ({ children }: { children: React$Node }) => {
         hasCompletedOnboarding &&
         wcContext.initDone &&
         !wcContext.session.session,
+      getStateFromPath: (path, config) => {
+        const url = new URL(`ledgerlive://${path}`);
+        const { hostname, pathname } = url;
+        const platform = pathname.split("/")[1];
+        if (hostname === "discover" && platform) {
+          /**
+           * Upstream validation of "ledgerlive://discover/:platform":
+           *  - checking that a manifest exists
+           *  - adding "name" search param
+           * */
+          const manifest = filteredManifests.find(
+            m => m.id.toLowerCase() === platform.toLowerCase(),
+          );
+          if (!manifest) return undefined;
+          url.pathname = `/${manifest.id}`;
+          url.searchParams.set("name", manifest.name);
+          return getStateFromPath(url.href?.split("://")[1], config);
+        }
+        return getStateFromPath(path, config);
+      },
     }),
-    [hasCompletedOnboarding, wcContext.initDone, wcContext.session.session],
+    [
+      hasCompletedOnboarding,
+      wcContext.initDone,
+      wcContext.session.session,
+      filteredManifests,
+    ],
   );
 
   const [isReady, setIsReady] = React.useState(false);
@@ -377,8 +432,24 @@ const DeepLinkingNavigator = ({ children }: { children: React$Node }) => {
     if (!wcContext.initDone) {
       return;
     }
+    if (
+      removeLiveAppState.isLoading &&
+      !removeLiveAppState.lastUpdateTime &&
+      !removeLiveAppState.error
+    )
+      /**
+       * Ensure that the list of manifests has been loaded once so that the
+       * deep linking logic to platform apps works in the scenario where the app
+       * was not previously running.
+       *  */
+      return;
     setIsReady(true);
-  }, [wcContext.initDone]);
+  }, [
+    wcContext.initDone,
+    removeLiveAppState.isLoading,
+    removeLiveAppState.lastUpdateTime,
+    removeLiveAppState.error,
+  ]);
 
   React.useEffect(
     () => () => {
@@ -470,6 +541,7 @@ export default class Root extends Component<
                 <SetEnvsFromSettings />
                 <HookSentry />
                 <AdjustProvider />
+                <DelayedTrackingProvider />
                 <HookAnalytics store={store} />
                 <WalletConnectProvider>
                   <RemoteLiveAppProvider