+++ apps/ledger-live-mobile/CHANGELOG.md
@@ -1,5 +1,55 @@
 # live-mobile
 
+## 3.3.0-next.5
+
+### Patch Changes
+
+- 3d71946fb: Fix deep linking logic for platform apps
+
+## 3.3.0-next.4
+
+### Minor Changes
+
+- 4db0f58ca: Updated learn feature url to be remote configurable
+
+## 3.3.0-next.3
+
+### Minor Changes
+
+- fe3c7a39c: Add Earn Rewards btn on Solana
+
+## 3.3.0-next.2
+
+### Minor Changes
+
+- 09f79c7b4: Add Cardano to mobile
+
+## 3.3.0-next.1
+
+### Minor Changes
+
+- 592ad2f7b: Update design on upsell modal in mobile app. Also add new variants and shape to IconBoxList and BoxedIcon components in native UI.
+- 608010c9d: Add a purchase device page embedding a webview from the ecommerce team. Also abstract webview pages logic into its own component (include Learn page's webview). Add a delayed tracking provider to send events to Adjust or Segment with an anonymised timestamp for sensible data.
+
+### Patch Changes
+
+- Updated dependencies [592ad2f7b]
+- Updated dependencies [608010c9d]
+  - @ledgerhq/native-ui@0.8.0-next.1
+  - @ledgerhq/live-common@23.1.0-next.1
+
+## 3.2.1-next.0
+
+### Patch Changes
+
+- 35737e057: fix(LLM): can't open custom loaded manifest on iOS [LIVE-2481]
+- 8a973ad0e: fix(LLM): platform manifest can be undefined when no network [LIVE-2571]
+- Updated dependencies [8323d2eaa]
+- Updated dependencies [8861c4fe0]
+- Updated dependencies [dd6a12c9b]
+  - @ledgerhq/live-common@23.1.0-next.0
+  - @ledgerhq/native-ui@0.7.19-next.0
+
 ## 3.2.0
 
 ### Minor Changes
@@ -7,6 +57,95 @@
 - becfc06f9: LIVE-1751 Solana staking on LLM
 - d63570a38: Rework Cosmos delegation flow
 - c6c127630: We now prompt a modal to ask the user what he thinks of the app at key moments (for example when receiving crypto or claiming rewards) based on some conditions (installed the app for at least x days, has at least x accounts, ...) The purpose of this feature is to increase the ratings of the app on the stores
+- 8ea9c2deb: LLM: increase to iOS 13 minimum
+- 64c2fdb06: Filecoin integration in LLD and LLM
+
+### Patch Changes
+
+- 09648db7f: refactor of the top perfs filter
+- 0f59cfc10: Fix crash related to the way polkadot/crypto handles environments that lack WASM support.
+- 68cb59649: Fix overlapped price on the market screen
+- 9a86fe231: Fix the click on browse assets button on the market screen
+- 54dbab04f: Fix Ledger logo glitch
+- 71ad84023: Track in Sentry the uncaught errors thrown in the bridge transaction flow.
+- 61116f39f: Disable React Navigation Sentry integration
+- Updated dependencies [09648db7f]
+- Updated dependencies [a66fbe852]
+- Updated dependencies [0f59cfc10]
+- Updated dependencies [8ee9c5568]
+- Updated dependencies [899aa3300]
+- Updated dependencies [89e82ed79]
+- Updated dependencies [403ea8efe]
+- Updated dependencies [98ecc6272]
+- Updated dependencies [9a86fe231]
+- Updated dependencies [8b2e24b6c]
+- Updated dependencies [64c2fdb06]
+- Updated dependencies [f686ec781]
+- Updated dependencies [b688a592d]
+- Updated dependencies [71ad84023]
+- Updated dependencies [64c2fdb06]
+  - @ledgerhq/live-common@23.0.0
+  - @ledgerhq/native-ui@0.7.18
+
+## 3.2.0-next.4
+
+### Patch Changes
+
+- Updated dependencies [8b2e24b6c]
+  - @ledgerhq/live-common@23.0.0-next.4
+
+## 3.2.0-next.3
+
+### Patch Changes
+
+- Updated dependencies [a66fbe852]
+  - @ledgerhq/live-common@23.0.0-next.3
+
+## 3.2.0-next.2
+
+### Patch Changes
+
+- Updated dependencies [8ee9c5568]
+  - @ledgerhq/live-common@23.0.0-next.2
+
+## 3.2.0-next.1
+
+### Patch Changes
+
+- Updated dependencies [98ecc6272]
+  - @ledgerhq/live-common@23.0.0-next.1
+
+## 3.2.0-next.0
+
+### Minor Changes
+
+- 8ea9c2deb: LLM: increase to iOS 13 minimum
+- 64c2fdb06: Filecoin integration in LLD and LLM
+
+### Patch Changes
+
+- 09648db7f: refactor of the top perfs filter
+- 0f59cfc10: Fix crash related to the way polkadot/crypto handles environments that lack WASM support.
+- 68cb59649: Fix overlapped price on the market screen
+- 9a86fe231: Fix the click on browse assets button on the market screen
+- 54dbab04f: Fix Ledger logo glitch
+- 71ad84023: Track in Sentry the uncaught errors thrown in the bridge transaction flow.
+- 61116f39f: Disable React Navigation Sentry integration
+- Updated dependencies [09648db7f]
+- Updated dependencies [0f59cfc10]
+- Updated dependencies [899aa3300]
+- Updated dependencies [89e82ed79]
+- Updated dependencies [403ea8efe]
+- Updated dependencies [9a86fe231]
+- Updated dependencies [64c2fdb06]
+- Updated dependencies [f686ec781]
+- Updated dependencies [b688a592d]
+- Updated dependencies [71ad84023]
+- Updated dependencies [64c2fdb06]
+  - @ledgerhq/live-common@23.0.0-next.0
+  - @ledgerhq/native-ui@0.7.18-next.0
+
+## 3.1.2
 
 ### Patch Changes